refactor: move tasks and management commands to main project app
Move Celery tasks and management commands from frontend app to main
parliscope project app for better separation of concerns:

- Move frontend/tasks.py → parliscope/tasks/indexing.py
- Move frontend/management/commands/update_solr.py → parliscope/management/commands/update_solr.py
- Move frontend/views.update → parliscope/views.update
- Update all import statements and URL routing
- Remove system orchestration code from user-facing frontend app

This creates cleaner architecture where:
- frontend/ handles user interface only
- parliscope/ handles system orchestration (tasks, commands, admin endpoints)

All functionality preserved, URLs unchanged (/update/ endpoint maintained).
Celery autodiscovery works with standard Django configuration.
